requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.30.0
psycopg2-binary>=2.9.9
instagrapi>=2.0.0
lxml>=5.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
